requirements.txt
# Server dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# Integration dependencies
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
notion-client==2.0.0
zoom-api-client==0.3.0

# Utility libraries
requests==2.31.0
python-jose==3.3.0
aiohttp==3.8.6
langchain==0.0.312
openai==1.2.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1 